Move functions used by only one command to the command file.
