+++ pkg/odo/cli/describe.go
@@ -1,9 +1,11 @@
 package cli
 
 import (
+	"fmt"
 	"github.com/redhat-developer/odo/pkg/component"
 	"github.com/redhat-developer/odo/pkg/odo/genericclioptions"
 	"github.com/redhat-developer/odo/pkg/odo/util"
+	"github.com/redhat-developer/odo/pkg/storage"
 	"github.com/spf13/cobra"
 )
 
@@ -45,4 +47,20 @@ func init() {
 	AddApplicationFlag(describeCmd)
 
 	rootCmd.AddCommand(describeCmd)
+}
+
+// printComponentInfo prints Component Information like path, URL & storage
+func printComponentInfo(currentComponentName string, componentType string, path string, componentURL string, appStore []storage.StorageInfo) {
+	// Source
+	if path != "" {
+		fmt.Println("Component", currentComponentName, "of type", componentType, "with source in", path)
+	}
+	// URL
+	if componentURL != "" {
+		fmt.Println("Externally exposed via", componentURL)
+	}
+	// Storage
+	for _, store := range appStore {
+		fmt.Println("Storage", store.Name, "of size", store.Size)
+	}
 }
